requirements.txt
# Core Framework
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
langchain-text-splitters>=0.3.0

# Vector Database & Embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Retrieval (BM25, Ensemble for hybrid)
rank_bm25>=0.2.2
langchain-classic>=1.0.0

# Utilities
python-dotenv>=1.0.1
wikipedia>=1.4.0
PyYAML>=6.0